Clean up the files and ready for Pull Request.
