+++ .gitignore
@@ -125,6 +125,3 @@ data/*txt
 output/
 *.zip
 *.p
-
-
-
